pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spotify-playlist-exporter"
version = "1.0.0"
description = "Export Spotify playlists to JSON, CSV, and XML formats"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["spotify", "playlist", "export", "json", "csv", "xml"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
spotify-playlist-exporter = "spotify_playlist_exporter.main:main"

[project.urls]
Homepage = "https://github.com/yourusername/spotify-playlist-exporter"
Documentation = "https://github.com/yourusername/spotify-playlist-exporter#readme"
Repository = "https://github.com/yourusername/spotify-playlist-exporter.git"
Issues = "https://github.com/yourusername/spotify-playlist-exporter/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --strict-markers"

[tool.coverage.run]
source = ["src/spotify_playlist_exporter"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
